pyproject.toml
[project]
name = "prism"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "arugo11", email = "al23088@shibaura-it.ac.jp" }
]
dependencies = [
    "pymongo[srv]>=4.10.1",
    "flask>=3.0.3",
    "python-dotenv>=1.0.1",
    "anthropic>=0.37.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/prism"]
